refactor: separate FloorPicture component
